Half-Baked Poral Construction
When finished, we can just call an instance of this class to do the pathfinding, selecting, and building of a nether portal site for us.

Unfortunately, placing blocks and liquid buckets in a structure that ISN'T set relative to a compass direction at write-time is REALLY hard due to place block needing a reference face.

That's the main thing that I think this will snag on first.