Constraint Improvements
While still inconsistent on the Z axis, the bot can now identify valid portal build sites that are parallel with the X Axis.
This site is returned as an array of arrays of blocks. The first (0) array is the lava, and the second (1) array is the coastline of ground blocks next to it. As long as this region (and the three blocks above it) is clear, a portal can be built there.